+++ src/mastodon_api/accounts/types.rs
@@ -446,12 +446,9 @@ pub struct IdentityProofData {
     pub created_at: DateTime<Utc>,
 }
 
-// TODO: actix currently doesn't support parameter arrays
-// https://github.com/actix/actix-web/issues/2044
 #[derive(Deserialize)]
 pub struct RelationshipQueryParams {
-    #[serde(rename(deserialize = "id[]"))]
-    pub id: Uuid,
+    pub id: Vec<Uuid>,
 }
 
 #[derive(Serialize)]